+++ tests/e2e/000-setup/010-create-tmp-dir.ts
@@ -1,9 +1,3 @@
-import {join} from 'path';
-
-import {isMobileTest, ng, expectFileToExist} from './utils';
-import {oneLine} from 'common-tags';
-
-
 const temp = require('temp');
