+++ tests/e2e/100-build/200-base-href.ts
@@ -1,7 +1,8 @@
-import {silentNg, fileMatchesOrFail} from '../utils';
+import {silentNg} from '../utils/process';
+import {expectFileToMatch} from '../utils/fs';
 
 
 export default function() {
   return silentNg('build', '--base-href', '/myUrl')
-    .then(() => fileMatchesOrFail('dist/index.html', /<base href="\/myUrl">/));
+    .then(() => expectFileToMatch('dist/index.html', /<base href="\/myUrl">/));
 }
