+++ tests/e2e/100-build/200-environment.ts
@@ -1,10 +1,13 @@
-import {silentNg, fileMatchesOrFail, expectGitToBeClean, expectToFail, ng} from '../utils';
+import {silentNg, ng} from '../utils/process';
+import {expectFileToMatch} from '../utils/fs';
+import {expectGitToBeClean} from '../utils/git';
+import {expectToFail} from '../utils/utils';
 
 
 export default function() {
   // Try a prod build.
   return silentNg('build', '--env=prod')
-    .then(() => fileMatchesOrFail('dist/main.bundle.js', 'production: true'))
+    .then(() => expectFileToMatch('dist/main.bundle.js', 'production: true'))
     .then(() => expectGitToBeClean())
 
     // Build fails on invalid build target
