+++ tests/e2e/100-build/200-output-dir.ts
@@ -1,6 +1,7 @@
-import {
-  silentNg, expectFileToExist, silentExecOrFail, expectGitToBeClean, expectToFail
-} from '../utils';
+import {silentNg, silentExecOrFail} from '../utils/process';
+import {expectFileToExist} from '../utils/fs';
+import {expectToFail} from '../utils/utils';
+import {expectGitToBeClean} from '../utils/git';
 
 
 export default function() {
